Fixed weights update in Gradient Descent.

+++ src/ActivationFunctions.java
@@ -2,35 +2,34 @@ public class ActivationFunctions {
 
     public ActivationFunctions(){}
 
-    protected float logistic(float x){
-        return (float) (1/(1+Math.exp(-x)));
+    protected double logistic(double x){
+        return 1 / (1+Math.exp(-x));
     }
 
-    protected float tanh(float x){
-        return (float) ((Math.exp( x - Math.exp(-x)))
-                / ( Math.exp( x + Math.exp(-x) ) ) );
+    protected double tanh(double x){
+        return (Math.exp( x - Math.exp(-x))) / (Math.exp( x + Math.exp(-x))) ;
     }
 
-    protected float relu(float x){
+    protected double relu(double x){
         return Math.max(0,x);
     }
 
-    protected float derivativeTanh(float x){
-        return (float) (1 - Math.pow(tanh(x), 2));
+    protected double derivativeTanh(double x){
+        return 1 - Math.pow(tanh(x), 2);
     }
 
-    protected float derivativeRelu(float x){
+    protected double derivativeRelu(double x){
         if (x < 0){
             return 0F;
         }
         return 1F;
     }
 
-    protected float derivativeLogistic(float x){
-        return (float) (Math.exp(-x)/Math.pow((1 + Math.exp(-x)), 2));
+    protected double derivativeLogistic(double x){
+        return Math.exp(-x) / Math.pow((1 + Math.exp(-x)), 2);
     }
 
-    public float selectActivationFunction(float x, String activationFunction){
+    public double selectActivationFunction(double x, String activationFunction){
         return switch (activationFunction) {
             case "tanh" -> tanh(x);
             case "relu" -> relu(x);
@@ -38,7 +37,7 @@ public class ActivationFunctions {
         };
     }
 
-    public float selectBackpropagationDerivative(float x, String activationFunction){
+    public double selectBackpropagationDerivative(double x, String activationFunction){
         return switch (activationFunction) {
             case "tanh" -> derivativeTanh(x);
             case "relu" -> derivativeRelu(x);
